TT-18010 preserve JSON-RPC codes as int64 end to end

+++ analytics/mcp_record.go
@@ -10,7 +10,7 @@ var MCPSQLTableName string
 // efficient querying while embedding the full AnalyticsRecord for all
 // standard analytics dimensions.
 type MCPRecord struct {
-	JSONRPCErrorCode         int    `json:"jsonrpc_error_code" bson:"jsonrpc_error_code" gorm:"column:jsonrpc_error_code"`
+	JSONRPCErrorCode         int64  `json:"jsonrpc_error_code" bson:"jsonrpc_error_code" gorm:"column:jsonrpc_error_code"`
 	JSONRPCMethod            string `gorm:"column:jsonrpc_method"`
 	PrimitiveType            string `gorm:"column:primitive_type"`
 	PrimitiveName            string `gorm:"column:primitive_name"`
